Move SoftwareCanvasProvider into its own source file (#30856)
