+++ ci/licenses_golden/licenses_flutter
@@ -40,6 +40,7 @@ FILE: ../../../flutter/display_list/display_list_benchmarks.cc
 FILE: ../../../flutter/display_list/display_list_benchmarks.h
 FILE: ../../../flutter/display_list/display_list_benchmarks_gl.cc
 FILE: ../../../flutter/display_list/display_list_benchmarks_metal.cc
+FILE: ../../../flutter/display_list/display_list_benchmarks_software.cc
 FILE: ../../../flutter/display_list/display_list_builder.cc
 FILE: ../../../flutter/display_list/display_list_builder.h
 FILE: ../../../flutter/display_list/display_list_canvas_dispatcher.cc
